Warning fixes.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@42490 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/fileback.cpp
@@ -83,7 +83,7 @@ wxBackingFileImpl::wxBackingFileImpl(wxInputStream *stream,
     wxFileOffset len = m_stream->GetLength();
 
     if (len >= 0 && len + size_t(1) < m_bufsize)
-        m_bufsize = len + 1;
+        m_bufsize = size_t(len + 1);
 
     if (m_bufsize)
         m_buf = new char[m_bufsize];
